adding PHPStan
