+++ composer.json
@@ -35,8 +35,10 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.3",
-        "phpstan/phpstan": "^0.9.0",
-        "phpunit/phpunit": "^6.0"
+        "phpunit/phpunit": "^6.0",
+        "phpstan/phpstan": "^0.9.2",
+        "phpstan/phpstan-strict-rules": "^0.9.0",
+        "phpstan/phpstan-phpunit": "^0.9.4"
     },
     "autoload": {
         "psr-4": {
@@ -50,11 +52,22 @@
         }
     },
     "scripts": {
+        "phpcs": "php-cs-fixer fix -v --diff --dry-run --allow-risky=yes;",
+        "phpstan-src": "phpstan analyse -l 7 -c phpstan.src.neon src",
+        "phpstan-tests": "phpstan analyse -l 7 -c phpstan.tests.neon tests",
+        "phpstan": [
+            "@phpstan-src",
+            "@phpstan-tests"
+        ],
+        "phpunit": "phpunit --coverage-text",
         "post-install-cmd": "\\League\\Uri\\Installer\\ICANNSection::update",
         "post-update-cmd": "\\League\\Uri\\Installer\\ICANNSection::update",
-        "test": "phpunit --coverage-text; php-cs-fixer fix -v --diff --dry-run --allow-risky=yes;",
-        "phpunit": "phpunit --coverage-text",
-        "phpcs": "php-cs-fixer fix -v --diff --dry-run --allow-risky=yes;"
+        "test": [
+            "@phpunit",
+            "@phpcs",
+            "@phpstan-src",
+            "@phpstan-tests"
+        ]
     },
     "extra": {
         "branch-alias": {
