+++ phpstan.src.neon
@@ -1,0 +1,2 @@
+includes:
+- vendor/phpstan/phpstan-strict-rules/rules.neon
